+++ CHANGELOG.md
@@ -7,6 +7,15 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 
 _No documentation available about unreleased changes as of yet._
 
+## [1.0.7] - 29.07.2019
+
+### Changed
+- Transfering helper render methods to class-block.php.
+- Changed get_blocks_path to non static method.
+
+### Removed
+- Removed class-block-view-helper.php.
+
 ## [1.0.6] - 25.07.2019
 
 ### Changed